+++ src/main/java/com/lukefowles/WeatherRestApiApplication.java
@@ -1,4 +1,4 @@
-package com.lukefowles.weather_rest_api;
+package com.lukefowles;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
