+++ src/main/java/com/lukefowles/exception/APIServerException.java
@@ -1,0 +1,7 @@
+package com.lukefowles.exception;
+
+public class APIServerException extends RuntimeException {
+    public APIServerException(String message) {
+        super(message);
+    }
+}
