+++ src/main/java/com/lukefowles/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package com.lukefowles.exception;
+
+public class BadRequestException extends RuntimeException {
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
