+++ src/main/java/com/lukefowles/exception/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.lukefowles.weather_rest_api;
+package com.lukefowles.exception;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
