+++ src/main/java/com/lukefowles/exception/LocationNotFoundException.java
@@ -1,0 +1,7 @@
+package com.lukefowles.exception;
+
+public class LocationNotFoundException extends RuntimeException {
+    public LocationNotFoundException(String message) {
+        super(message);
+    }
+}
