+++ src/main/java/com/lukefowles/weather/WeatherApiCallRepository.java
@@ -1,4 +1,4 @@
-package com.lukefowles.weather_rest_api;
+package com.lukefowles.weather;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-interface WeatherApiCallRepository extends JpaRepository<OpenWeatherApiCall, Long> {
+public interface WeatherApiCallRepository extends JpaRepository<OpenWeatherApiCall, Long> {
     List<OpenWeatherApiCall> findByLocation(String location);
 }
