+++ src/main/java/com/lukefowles/weather/WeatherController.java
@@ -1,4 +1,4 @@
-package com.lukefowles.weather_rest_api;
+package com.lukefowles.weather;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
