+++ src/main/java/com/lukefowles/weather/WeatherResponse.java
@@ -1,4 +1,4 @@
-package com.lukefowles.weather_rest_api;
+package com.lukefowles.weather;
 
 import java.util.Objects;
