--- src/main/java/com/lukefowles/weather_rest_api/APIServerException.java
@@ -1,7 +1,0 @@
-package com.lukefowles.weather_rest_api;
-
-class APIServerException extends RuntimeException {
-    APIServerException(String message) {
-        super(message);
-    }
-}
