--- src/main/java/com/lukefowles/weather_rest_api/BadRequestException.java
@@ -1,7 +1,0 @@
-package com.lukefowles.weather_rest_api;
-
-class BadRequestException extends RuntimeException {
-    BadRequestException(String message) {
-        super(message);
-    }
-}
