--- src/main/java/com/lukefowles/weather_rest_api/LocationNotFoundException.java
@@ -1,7 +1,0 @@
-package com.lukefowles.weather_rest_api;
-
-class LocationNotFoundException extends RuntimeException {
-    LocationNotFoundException(String message) {
-        super(message);
-    }
-}
