+++ src/test/java/com/lukefowles/weather/WeatherServiceTests.java
@@ -1,13 +1,11 @@
-package com.lukefowles.weather_rest_api;
+package com.lukefowles.weather;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.*;
 
-import java.io.IOException;
 import java.time.Instant;
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.Mockito.*;
